fixed employee form

+++ SafeBoxes_TS2_ALAFAK/Frm_NewEmployee.vb
@@ -1,6 +1,8 @@
 ﻿Public Class Frm_NewEmployee
     Dim theNewId As Integer
+    Dim isSubmitting As Boolean = False
     Private Sub btn_submit_Click(sender As Object, e As EventArgs) Handles btn_submit.Click
+        isSubmitting = True
         If txt_fname.Text = "" Or txt_lname.Text = "" Or txt_father.Text = "" Then
             MessageBox.Show("Please fill all needed information!")
         Else
@@ -11,14 +13,16 @@
     End Sub
 
     Private Sub Frm_NewEmployee_Closed(sender As Object, e As EventArgs) Handles Me.Closed
-        ExecuteQuery("DELETE FROM Employees WHERE EmpId = " & theNewId)
-        Me.Dispose()
+        If isSubmitting = False Then
+            ExecuteQuery("DELETE FROM Employees WHERE EmpId = " & theNewId)
+            Me.Dispose()
+        End If
     End Sub
 
     Private Sub Frm_NewEmployee_Load(sender As Object, e As EventArgs) Handles MyBase.Load
         theNewId = genID("Employees", "EmpId")
         ExecuteQuery("INSERT INTO Employees(EmpId) VALUES(" & theNewId & ")")
-        lbl_Empid.Text = "EmployeeId ID: " & theNewId
+        lbl_Empid.Text = "Employee's ID: " & theNewId
     End Sub
 
     Private Sub txt_father_KeyPress(sender As Object, e As KeyPressEventArgs) Handles txt_father.KeyPress
